Correct errors found by the nextgen CI work

+++ shakenfist/client/apiclient.py
@@ -278,11 +278,15 @@ class Client(object):
         return r.json()
 
     def delete_namespace(self, namespace):
-        # Ensure the namespace is empty first
-        for i in self.get_instances():
-            self.delete_instance(i['uuid'])
-        for n in self.get_networks():
-            self.delete_network(n['uuid'])
+        if not namespace:
+            namespace = self.namespace
+
+        if namespace != 'system':
+            # Ensure the namespace is empty first
+            for i in self.get_instances(all=True):
+                self.delete_instance(i['uuid'])
+            for n in self.get_networks():
+                self.delete_network(n['uuid'])
 
         self._request_url(
             'DELETE', self.base_url + '/auth/namespace/' + namespace)
